Added configuration processing, nproc cmdline arg, and genmap call.
The script can apply relations like:
number of elements = (number of elements in x) * (" " in y) * (" " in z)
This is done by loading the configuration dict into the current
variable scope, manipulating by parameter name, and then dumping
the current variable scope back into the configuration dict.

The number of processors can be specified on the command line, which
overrides the value in the configuration.

Added a call to genmap.